+++ app/src/main/java/com/nextcloud/utils/extensions/ActivityExtensions.kt
@@ -7,9 +7,9 @@
 
 package com.nextcloud.utils.extensions
 
-import android.app.Activity
-import android.os.Build
-import android.view.WindowInsets
+import android.view.View
+import android.view.Window
+import androidx.annotation.ColorInt
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
@@ -23,23 +23,22 @@ fun AppCompatActivity.fragments(): List<Fragment> = supportFragmentManager.fragm
 
 fun AppCompatActivity.lastFragment(): Fragment = fragments().last()
 
-fun Activity.hasNavigationBar(): Boolean {
-    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-        val insets = window.decorView.rootWindowInsets
-        insets?.isVisible(WindowInsets.Type.navigationBars()) == true
-    } else {
-        val insets = ViewCompat.getRootWindowInsets(window.decorView)
-        insets?.isVisible(WindowInsetsCompat.Type.navigationBars()) == true
+// TODO move it to the WindowExtensions
+fun Window?.setNavBarColor(@ColorInt color: Int) {
+    if (this == null) {
+        return
     }
-}
 
-fun Activity.navBarHeight(windowInsetsCompat: WindowInsetsCompat): Int {
-    val typeMask = WindowInsetsCompat.Type.navigationBars()
-    val insets = windowInsetsCompat.getInsets(typeMask)
-
-    return if (insets.bottom != 0 && hasNavigationBar()) {
-        insets.bottom
-    } else {
-        0
+    ViewCompat.setOnApplyWindowInsetsListener(decorView) { v: View, insets: WindowInsetsCompat ->
+        val navigationBarInsets = insets.getInsets(WindowInsetsCompat.Type.navigationBars())
+        v.setPadding(
+            v.paddingLeft,
+            v.top,
+            v.paddingRight,
+            navigationBarInsets.bottom,
+        )
+        insets
     }
+
+    decorView.setBackgroundColor(color)
 }
